Extract the attestation protocol into a transport-generic module
Factor the post-handshake attestation exchange out of client.rs and
server.rs into a new attest module, in preparation for running the same
protocol over a QUIC bidirectional stream as well as TCP.

- Add tls/src/attest.rs holding the framing (send_msg/recv_msg with its
  MAX_MSG_SIZE bound), the DER cert-chain helpers
  (certs_to_der/certs_from_der), the protocol version constants and result
  aliases, and corims_from_paths, all moved from lib.rs and server.rs.
- Extract client_exchange and server_exchange, each generic over
  T: AsyncRead + AsyncWrite + Unpin. They take the peer PlatformId already
  derived from the TLS handshake certificates and return the peer's
  attested PlatformId plus whether its measurements appraised. The bodies
  are the former inline exchanges verbatim, including the checked
  version-message parse and preserving the wire-visible asymmetries: the
  server sends its own attestation last (and never, under Enforced, after
  a failed appraisal), the client loads its corims inline at the end of
  the exchange, and the server loads its corpus before accepting the TLS
  connection so a malformed corpus fails early.
- Client::connect_with_config and SprocketsAcceptor::handshake shrink to
  TLS setup, peer-PlatformId extraction, and a call into attest.

No change to the attestation protocol bytes, the TLS configuration, or the
crypto provider. Verified on illumos: clippy --all-targets clean and the
TCP test suite passes.
